requirements.txt
# Untuk PostgresOperator
apache-airflow-providers-postgres

# Untuk fungsi transform_load (Pandas dan koneksi SQL)
pandas
sqlalchemy
psycopg2-binary